Parametric grammars, with checking and inference
Problem statement
-----------------

Often we would like our symbols to have extra type parameters. A static
name, for instance

  data Name (s :: Symbol) (t :: *) = Name t

To use this in a grammar, barring any use of the apparently
broken ImpredicativeTypes extension, requires introducing type
parameters to the grammar at large.

  type ExampleGrammar name1 name2 = GAllOf '[
        GSymbol (Name name1)
      , GOptional (GSymbol (Name name2))
      ]

Now we have an infinite family of grammars, one for each choice of
those parameters. But this seems to be awfully restrictive; before we
can parse type using that grammar, we have to specify exactly which type
parameters we're after. It would be much better if we had the option to
infer them from the type given to the parser!

Now we can :)

Just set the parameter to Infer if you want to infer, or Check s if you
want to restrict it to a certain type.

These changes bring an important caveat: the type that you get out of
the parser is not necessarily the same type that you ask it to parse for
you. It will be the same if and only if the grammar type contains no
parametric symbols.