Redesign: remove universal Grammar GADT
The idea is now to define *derived symbolic grammars* and for each of
them an associated *inferred grammar*. You specify the form via the
symbolic grammar, and after a parse, you get the inferred grammar, which
unlike the symbolic grammar is an inhabited type (except for
GEmpty/PEmpty, which never parses).

The intention is to give a nice high-level representation of the parsed
datatype, because working with only the primitives like GProduct and
GSum would be very annoying. Observe the example non-trivial derived
grammars GOneOf, GAllOf, GMany. The form is like so: declare the grammar
from which it is derived, by giving a DerivedGrammar instance and
associated DerivedFrom type. Then, define how the inferred type is
produced from the inferred type of the grammar from which it is derived,
by giving an instance of InferredGrammar.