Generalize the infix grammar
Now we can have prefixes and postfixes. For associative variants, only a
prefix or a postfix is allowed (prefix for right, postfix for left).

Motivation: we sometimes encounter these associative infix grammars in
which there is a prefix or a postfix, for example in the SQL join
syntax. JOIN is left-associative, but also has a postfix ON or USING
clause.