Check in grammar WIP for historical value
I was shocked to discover that the infix grammar called GradeSchool
in Precedence.hs took at least minutes and 4 gigabytes of memory to
compile (I never actually saw the end of it). I did some investigation
and inferred that GHC was choking because of all of the repeated
ParseGrammarK and TransitiveInferredGrammarRec calls. I cleaned it up
and saw a massive improvement but concluded that it's still too damn
slow to be of any use at all. Even a simple GSepBy grammar would take
seconds to check. The cause of all this is an explosion in the number of
coercions in the Core generated from code using type grammar parsing.

So this work is abandoned, for now. I have come up with something that
looks more promising: faster and in direct analogy to applicative term
parsers.